+++ changelog.rst
@@ -3,6 +3,10 @@
 Change Log
 ##########
 
+- In development.
+
+  - Fix bug inserting bookmarks into empty files.
+
 - Version 0.1 (2022-01-02)
 
   Initial release.